Enable DMA for background transfer of I2S samples to the PIO
Previously, all samples were directly written to the PIO FIFO buffer in
a tight loop. This caused spurious clicks and sound artifacts and is not
viable in the long run. This commit explores the idea of using DMA to
directly stream pre-computed samples to the PIO. Since the HAL is not
that developed yet, the configuration has to be done through direct
register manipulations.