Finalize and clean up double buffering
However, there is a problem when one tries to step through the code.
The DMA streams keep triggering each other without the interrupt code,
that reloads the read address register of each stream, being executed.
This causes the DMA to read past the buffer end and write garbage to
the I2C which in turn yields a short burst of high volume noise.