Add Biquad filters (#3)
* Add biquad filter implementation and sawtooth waveform

Made the `SineWave` more general so that different lookup tables can be
used. Some debug prints were added to monitor the time it takes to
generate the samples each time. Seems the floating point operations are
too expensive as it is right now, and since the DMA channels are
configured to automatically trigger the other one when done, there is a
case when they start sending arbitrary memory contents to the amplifier
causing loud noise if we are not able to reset the read address on time.

* Update README